+++ src/components/Header.js
@@ -50,13 +50,13 @@ function Header() {
         >
           {state ?
           <svg xmlns="http://www.w3.org/2000/svg" width="18" height="14" viewBox="0 0 18 14" fill="none">
-            <path d="M17 1L1 1" stroke="#fff" stroke-width="1.5" stroke-linecap="round"/>
-            <path d="M17 7L1 7" stroke="#fff" stroke-width="1.5" stroke-linecap="round"/>
-            <path d="M17 13L1 13" stroke="#fff" stroke-width="1.5" stroke-linecap="round"/>
+            <path d="M17 1L1 1" stroke="#fff" strokeWidth="1.5" strokeLinecap="round"/>
+            <path d="M17 7L1 7" stroke="#fff" strokeWidth="1.5" strokeLinecap="round"/>
+            <path d="M17 13L1 13" stroke="#fff" strokeWidth="1.5" strokeLinecap="round"/>
           </svg>
           :
           <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-            <path d="M6.11936 6L18 18M17.8806 6L6 18" stroke="#fff" stroke-width="1.5" stroke-linecap="round"/>
+            <path d="M6.11936 6L18 18M17.8806 6L6 18" stroke="#fff" strokeWidth="1.5" strokeLinecap="round"/>
           </svg>
           }
         </button>
