+++ src/pages/LocationsList.js
@@ -1,10 +1,5 @@
 import ErrorBoundary from "../components/ErrorBoundary";
-import { useGetList } from "../hooks/useGetList";
-
-function List() {
-  const list = useGetList("https://rickandmortyapi.com/api/location", "locations");
-  return list;
-}
+import { LocationList } from "../components/Location/LocationList";
 
 function LocationsList() {
   return (
@@ -12,7 +7,7 @@ function LocationsList() {
       <div className="main-container pt-[72px]">
         <h1 className="page-title mb-6">Locations</h1>
         <ErrorBoundary>
-          <List />
+          <LocationList />
         </ErrorBoundary>
       </div>
     </div>
